mlops-pipeline/requirements.txt
# Core ML libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Version control & Experiment tracking
dvc==3.35.0
dvc-gdrive==2.20.0
mlflow==2.9.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# CI/CD (for local testing)
pyyaml==6.0.1

# Drift monitoring
evidently==0.4.18

# Utilities
python-dotenv==1.0.0
requests==2.31.0